Minor optimization

+++ pantaray/camera.hpp
@@ -8,7 +8,7 @@ namespace PantaRay {
 
         virtual void Begin() = 0;
 
-        virtual Ray GetRay(double x_interpolation, double y_interpolation) = 0;
+        virtual Ray GetRay(float x_interpolation, float y_interpolation) = 0;
 
         virtual ~ICamera() {}
 
@@ -25,11 +25,11 @@ namespace PantaRay {
         Vector target;
         Vector up;
 
-        double fov;
-        double aspect_ratio;
+        float fov;
+        float aspect_ratio;
 
     public:
-        PinholeCamera(double _fov, double _aspect_ratio) : fov(_fov), aspect_ratio(_aspect_ratio) {}
+        PinholeCamera(float _fov, float _aspect_ratio) : fov(_fov), aspect_ratio(_aspect_ratio) {}
 
         PinholeCamera& Set(Vector _position, Vector _target, Vector _up = Vector(0, 0, 1)) {
             position = _position;
@@ -44,7 +44,7 @@ namespace PantaRay {
             auto forward = target.Copy().Subtract(position).Normalize();
             auto left = up.Copy().Cross(forward).Normalize();
 
-            auto left_factor = tan(fov / 2);
+            auto left_factor = tanf(fov / 2);
             auto up_factor = left_factor / aspect_ratio;
 
             auto middle_left = left.Copy().Scale(left_factor);
@@ -55,7 +55,7 @@ namespace PantaRay {
             bottom_left = middle_left.Copy().Add(up.Copy().Scale(-up_factor)).Add(forward);
         }
 
-        Ray GetRay(double x_interpolation, double y_interpolation) {
+        Ray GetRay(float x_interpolation, float y_interpolation) {
             return Ray(
                 position.Copy(),
                 top_left.Copy()
