+++ pantaray/ray.hpp
@@ -11,7 +11,7 @@ namespace PantaRay {
 
         Ray(Vector& _start, Vector& _direction) : start(_start), direction(_direction) {}
 
-        Vector ScaleTo(double factor) const {
+        Vector ScaleTo(float factor) const {
             return start.Copy().Add(direction.Copy().Scale(factor));
         }
